Prevent race condition when cancelling long-press for scroll.
When dragging between workspaces in rapid succession, the canceling of the
original long-press timer wasn't being triggered correctly.  (When the timer
fires, it might read an invalid Workspace.allowLongPress() value.)

This patchset correctly cancels any pending long-press timers once a desktop
scroll begins, and we don't need to rely on the allowLongPress() value.