Passion - Maybe this will fix bug 2183130 - NPE in AllAppsView.onTouchEvent.
Looking at the code as it was on ERD11, this was mVelocity being null.
That shouldn't be possible if events are ordered correctly.  This
change makes AllAppsView a little more rigorous about which value
mTouchTracking has.  Hopefully, we can mask any underlying problems
sufficiently.  Events aren't supposed to be going out of order.