Fix bug 2242465 - Can add shortcuts while the "3d all apps" is open.
The more I think about it, the more I think AllApps should just be a separate
activity.  It's not a drawer anymore and you open it by pressing a button.
Also, it will want to derive from RenderScriptActivity or something, and
have all that stuff dealt with automatically too.  It's to late to make a
change like this now, but if we do choose to do it someday, I think we'll want
to reconsider the menu that's shown.

For now, just close all apps when they add a shortcut or change the wallpaper.