Launcher2 - This might fix the missing app problem and the crash in removeApp.
This marks the apps list as dirty by default.  Usually, because of how long
it takes the loader thread to run, the apps list comes in after the surface
has been initialized.  However, if the surface isn't around yet when setApps
is called, then setApps won't set mRollo.mAppsDirty because mRollo is null.
The extra null check is necessary for the normal case when mAllAppsList isn't
ready yet when the surface comes in.