Fix 2325492 - No icons in launcher after pressing home from within an app
It looks like the evil hack in 14f122bf847e50a3e7730ccbe57abc25d086a01b to make the workspace
not animate didn't completely work.  The key to reproducing this bug is to make sure the
activity is destroyed and to have last gone to an app from a screen other than the center
screen, because that causes it to get reloaded from the icicle, which makes the timing
more amenable to missing the animation, because the view isn't attached yet.