Widen AllApps3D text to match Launcher
The AllApps3D text used to have a bubble shade for the application
title to make it more visible over a translucent background of the
desktop.  This was later changed to a black background eliminating
the need for the bubble surrounding the text.  The majority of the
bubble rectangle code was commented out leaving all of the padding
still in effect. This caused application titles to wrap in varying
places between Launcher and Launcher2. Padding changes solve this.

Change-Id: Id67649562045f65b44409dfb81e6112a1f838143