Allow shared user id and process to be set via overlays.
This fiddles the manifest to allow the app's shared user id and
process to be changed at build time via resource overlays, so we
can build launcher to run in the acore process for low-end devices.

Also switches the cert of launcher back to shared, to allow for
this.  (Note this only really matters if the shared user id is set,
otherwise launcher will still get its own uid.)

Change-Id: I0bd0a06e1442c4bfe5ceb2d438cb58f29d911578