Fix for a Launcher crash.
A forced close will occur in Launcher homescreen if you
are holding the edge portion of the touch screen and slide
close or open the touch screen. This is caused by the
display changing configurations but the touch events are
passed through the framework. So the coordinates will be
wrong which causes a out of bounds exception.

Added a bounds check for the pointer index on ACTION_UP.

Change-Id: I67d7e428f55694969d43948c37b6acda2a5c6eb6