simplifying code to prevent drawing off-screen
On certain graphics drivers, if you draw to a off-
screen buffer that's not used, it can lead to
poor performance. We were running into this in 
launcher; when setChildrenLayersEnabled was called
on a CellLayout, that triggered a re-draw of that
CellLayout's hardware layer, even if that
CellLayout wasn't visible. This CL changes
PagedView so that pages that aren't going to be
rendered are set to View.INVISIBLE, preventing
re-drawing of their hardware layer.