Add support for uri favorites and resolve groups
Adds support for new types in default_workspace xml files.

Adds support to favorite that attempts to resolve an intent specified
by URI. If a logical choice can be found from the resolve, it adds the
shortcut.

Adds support for a resolve, which contains several favorites and attempts
to add the shortcuts until one is successful.

Bug: 10547358
Change-Id: Ia30341e77e7d4a0bd4a4f4e6a3d9ce2a48a46237